Add three more bls test cases and refactor scripts to separate modules
